Rename `LeanCode.DomainModels.MassTransitRelay` -> `LeanCode.CQRS.MassTransitRelay`
