+++ LeanCode.CoreLibrary.sln
@@ -107,9 +107,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.Serialization.Test
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.CQRS.RemoteHttp.Client.Tests", "test\Domain\LeanCode.CQRS.RemoteHttp.Client.Tests\LeanCode.CQRS.RemoteHttp.Client.Tests.csproj", "{E6A79F5B-8A30-4496-8F70-871C43AF91D4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.DomainModels.MassTransitRelay", "src\Domain\LeanCode.DomainModels.MassTransitRelay\LeanCode.DomainModels.MassTransitRelay.csproj", "{4E6FB2E6-FC2C-4814-BA80-0C917516896F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.CQRS.MassTransitRelay", "src\Domain\LeanCode.CQRS.MassTransitRelay\LeanCode.CQRS.MassTransitRelay.csproj", "{4E6FB2E6-FC2C-4814-BA80-0C917516896F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.DomainModels.MassTransitRelay.Tests", "test\Domain\LeanCode.DomainModels.MassTransitRelay.Tests\LeanCode.DomainModels.MassTransitRelay.Tests.csproj", "{80A33FE3-868B-49DE-9C78-E353204C3B2F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.CQRS.MassTransitRelay.Tests", "test\Domain\LeanCode.CQRS.MassTransitRelay.Tests\LeanCode.CQRS.MassTransitRelay.Tests.csproj", "{80A33FE3-868B-49DE-9C78-E353204C3B2F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeanCode.SendGrid", "src\Infrastructure\LeanCode.SendGrid\LeanCode.SendGrid.csproj", "{57BFC47F-7010-4FDB-92A8-5FF73D136F14}"
 EndProject
