+++ src/Domain/LeanCode.CQRS.MassTransitRelay/LeanCode.CQRS.MassTransitRelay.csproj
