--- test/Domain/LeanCode.DomainModels.MassTransitRelay.Tests/Integration/TestEvents.cs
@@ -1,47 +1,0 @@
-using System;
-using LeanCode.DomainModels.Model;
-using LeanCode.Time;
-
-namespace LeanCode.DomainModels.MassTransitRelay.Tests.Integration;
-
-public class Event1 : IDomainEvent
-{
-    public DateTime DateOccurred { get; }
-    public Guid Id { get; }
-    public Guid CorrelationId { get; }
-
-    public Event1(Guid correlationId)
-    {
-        CorrelationId = correlationId;
-        Id = Guid.NewGuid();
-        DateOccurred = TimeProvider.Now;
-    }
-}
-
-public class Event2 : IDomainEvent
-{
-    public DateTime DateOccurred { get; }
-    public Guid Id { get; }
-    public Guid CorrelationId { get; }
-
-    public Event2(Guid correlationId)
-    {
-        Id = Guid.NewGuid();
-        DateOccurred = TimeProvider.Now;
-        CorrelationId = correlationId;
-    }
-}
-
-public class Event3 : IDomainEvent
-{
-    public DateTime DateOccurred { get; }
-    public Guid Id { get; }
-    public Guid CorrelationId { get; }
-
-    public Event3(Guid correlationId)
-    {
-        Id = Guid.NewGuid();
-        DateOccurred = TimeProvider.Now;
-        CorrelationId = correlationId;
-    }
-}
