--- test/Domain/LeanCode.DomainModels.MassTransitRelay.Tests/LeanCode.DomainModels.MassTransitRelay.Tests.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="../../../src/Domain/LeanCode.DomainModels.MassTransitRelay/LeanCode.DomainModels.MassTransitRelay.csproj" />
-    <ProjectReference Include="../../../src/Domain/LeanCode.TimeProvider/LeanCode.TimeProvider.csproj" />
-    <ProjectReference Include="../../../src/Testing/LeanCode.IntegrationTestHelpers/LeanCode.IntegrationTestHelpers.csproj" />
-
-    <ProjectReference Include="../../LeanCode.Test.Helpers/LeanCode.Test.Helpers.csproj" />
-
-    <ProjectReference Include="../../../src/Domain/LeanCode.CQRS.Execution/LeanCode.CQRS.Execution.csproj" />
-    <ProjectReference Include="../../../src/Domain/LeanCode.CQRS.AspNetCore/LeanCode.CQRS.AspNetCore.csproj" />
-
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
-    <PackageReference Include="MassTransit.EntityFrameworkCore" />
-  </ItemGroup>
-
-</Project>
